Convert the keyboard module into a general controller module - still messy, but progressing
